+++ src/components/DataFetcher/ErrorMessage.tsx
@@ -4,17 +4,18 @@ import RetryIcon from '../../../public/icons/retry.svg';
 
 import styles from './ErrorMessage.module.scss';
 
+import { OFFLINE_ERROR } from 'src/api';
 import Button, { ButtonSize, ButtonType } from 'src/components/dls/Button/Button';
 
 interface Props {
   onRetryClicked: () => void;
-  isError?: boolean;
+  error: Error;
 }
 
-const ErrorMessage: React.FC<Props> = ({ onRetryClicked, isError = true }) => (
+const ErrorMessage: React.FC<Props> = ({ onRetryClicked, error }) => (
   <div className={styles.container}>
     <p className={styles.text}>
-      {isError
+      {error.message !== OFFLINE_ERROR
         ? 'Something went wrong. Please try again.'
         : 'Looks like you lost your connection. Please check it and try again.'}
     </p>
